requirements.txt
alembic==1.4.3
appdirs==1.4.4
argon2-cffi==20.1.0
asn1crypto==0.24.0
astroid==2.1.0
asttokens==1.1.13
async-generator==1.10
attrs==20.3.0
automationhat==0.2.2
backcall==0.2.0
beautifulsoup4==4.7.1
bleach==3.2.1
blinker==1.4
blinkt==0.1.2
buttonshim==0.0.2
Cap1xxx==0.1.3
certifi==2018.8.24
certipy==0.1.3
cffi==1.14.4
chardet==3.0.4
Click==7.0
colorama==0.3.7
colorzero==1.1
cookies==2.2.1
cryptography==2.6.1
cycler==0.10.0
decorator==4.3.0
defusedxml==0.6.0
distlib==0.3.1
docutils==0.14
drumhat==0.1.0
entrypoints==0.3
envirophat==1.0.0
ExplorerHAT==0.4.2
filelock==3.0.12
Flask==1.0.2
fourletterphat==0.1.0
gpiozero==1.5.1
h5py==2.8.0
html5lib==1.0.1
idna==2.6
imageio==2.9.0
importlib-metadata==3.1.0
imutils==0.5.3
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.4
itsdangerous==0.24
jedi==0.13.2
Jinja2==2.10
jsonschema==3.2.0
keyring==17.1.1
keyrings.alt==3.1.1
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
logilab-common==1.4.2
lxml==4.3.2
Mako==1.1.3
MarkupSafe==1.1.0
matplotlib==3.0.2
mccabe==0.6.1
microdotphat==0.2.1
mistune==0.8.4
mote==0.0.4
motephat==0.0.3
mypy==0.670
mypy-extensions==0.4.1
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
notebook==6.1.5
numpy==1.16.2
oauthlib==3.1.0
olefile==0.46
packaging==20.4
pamela==1.0.0
pandocfilters==1.4.3
pantilthat==0.0.7
parso==0.3.1
pexpect==4.8.0
pgzero==1.2
phatbeat==0.1.1
pianohat==0.1.0
picamera==1.13
pickleshare==0.7.5
piglow==1.2.5
pigpio==1.44
Pillow==5.4.1
prometheus-client==0.9.0
prompt-toolkit==3.0.8
psutil==5.5.1
ptyprocess==0.6.0
pycparser==2.20
pycrypto==2.6.1
pygame==1.9.4.post1
Pygments==2.7.2
PyGObject==3.30.4
pyinotify==0.9.6
PyJWT==1.7.0
pylint==2.2.2
pyOpenSSL==19.0.0
pyparsing==2.2.0
pypylon==1.6.0
pypylon-opencv-viewer==1.1.0
pyrsistent==0.17.3
pyserial==3.4
python-apt==1.8.4.1
python-dateutil==2.7.3
python-editor==1.0.4
python-json-logger==2.0.1
pyxdg==0.25
pyzmq==20.0.0
qtconsole==5.0.1
QtPy==1.9.0
rainbowhat==0.1.0
requests==2.21.0
requests-oauthlib==1.0.0
responses==0.9.0
roman==2.0.0
RPi.GPIO==0.7.0
RTIMULib==7.2.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
scipy==1.1.0
scrollphat==0.0.7
scrollphathd==1.2.1
SecretStorage==2.3.1
Send2Trash==1.5.0
simplejson==3.16.0
six==1.12.0
skywriter==0.0.7
sn3218==1.2.7
soupsieve==1.8
spidev==3.4
SQLAlchemy==1.3.20
ssh-import-id==5.7
ST7735==0.0.4
terminado==0.9.1
testpath==0.4.4
thonny==3.3.0
tornado==6.1
touchphat==0.0.1
traitlets==5.0.5
twython==3.7.0
typed-ast==1.3.1
unicornhathd==0.0.4
urllib3==1.24.1
virtualenv==20.2.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.5.1
wrapt==1.10.11
ws4py==0.5.1
zipp==3.4.0
